Final changes GA used for all results

+++ GA/Results.cs
@@ -8,14 +8,14 @@ namespace GA
 {
     class Results
     {
-        public List<int> GenerationsTaken { get; set; }
+        public List<float> GenerationsTaken { get; set; }
         public List<String> GenerationResults { get; set; }
         public String Parameters { get; set; }
         public int numOfDecptiveBits { get; set; }
 
         public Results()
         {
-            GenerationsTaken = new List<int>();
+            GenerationsTaken = new List<float>();
             GenerationResults = new List<String>();
             Parameters = "";
             numOfDecptiveBits = 0;
